requirements.txt
-i https://pypi.org/simple
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
black==18.9b0
click==7.0
cycler==0.10.0
flask==1.0.2
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl#egg=torch
itsdangerous==1.1.0
jinja2==2.10
kiwisolver==1.0.1
markupsafe==1.1.1
matplotlib==3.0.3
more-itertools==6.0.0 ; python_version > '2.7'
numpy==1.16.2
opencv-contrib-python==4.0.0.21
pillow==5.4.1
pluggy==0.9.0
py==1.8.0
pyparsing==2.3.1
pytest==4.3.1
python-dateutil==2.8.0
six==1.12.0
toml==0.10.0
torchvision==0.2.2.post3
werkzeug==0.15.0
